+++ sdk/resourcemanager/servicelinker/armservicelinker/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Release History
 
+## 0.5.0 (2022-05-09)
+### Breaking Changes
+
+- Type of `AzureResourcePropertiesBase.Type` has been changed from `*Type` to `*AzureResourceType`
+- Type of `TargetServiceBase.Type` has been changed from `*Type` to `*TargetServiceType`
+- Type of `AzureResource.Type` has been changed from `*Type` to `*TargetServiceType`
+- Type of `ConfluentBootstrapServer.Type` has been changed from `*Type` to `*TargetServiceType`
+- Type of `ConfluentSchemaRegistry.Type` has been changed from `*Type` to `*TargetServiceType`
+- Type of `AzureKeyVaultProperties.Type` has been changed from `*Type` to `*AzureResourceType`
+- Const `TypeAzureResource` has been removed
+- Const `TypeConfluentBootstrapServer` has been removed
+- Const `TypeConfluentSchemaRegistry` has been removed
+- Const `TypeKeyVault` has been removed
+- Function `PossibleTypeValues` has been removed
+- Field `ValidateResult` of struct `LinkerClientValidateResponse` has been removed
+
+### Features Added
+
+- New const `TargetServiceTypeConfluentSchemaRegistry`
+- New const `TargetServiceTypeConfluentBootstrapServer`
+- New const `AzureResourceTypeKeyVault`
+- New const `TargetServiceTypeAzureResource`
+- New function `PossibleAzureResourceTypeValues() []AzureResourceType`
+- New function `PossibleTargetServiceTypeValues() []TargetServiceType`
+- New struct `ValidateOperationResult`
+- New anonymous field `ValidateOperationResult` in struct `LinkerClientValidateResponse`
+
+
 ## 0.4.0 (2022-04-18)
 ### Breaking Changes